This was used by the streamer spec to assert linear complexity, but causes a huge amount of array allocations with 300 MB memory usage and takes up about 40% of the entire test suite runtime for a single spec.

Instead replace it with a simple inline implementation that asserts linear complexity with 75% confidence interval.

While local tests show that the usual variance is within 5% or 95% confidence, the lower confidence is needed to account for fluctuation between test runs caused by GC or system load.

This could be counted by going for higher file counts, but starting from about 10K files that makes the spec unbearably slow.