@felixbuenemann 2020-11-24
Argh, I rearranged the blocks and am measuring file size instead of execution time, will fix.