@felixbuenemann 2020-11-25
@julik Another approach would be to tag the slow spec and only run it on demand.

The code for this ended up much more complex that my initial PoC to avoid flipping due to variance.