@felixbuenemann 2020-11-25
### Performance comparison

RAM: 468M -> 33M (14x reduction)
CPU: 7.19s -> 0.56s (13x reduction)

**Before:**

```
$ /usr/bin/time -l rspec spec/zip_tricks/streamer_spec.rb:31
Run options: include {:locations=>{"./spec/zip_tricks/streamer_spec.rb"=>[31]}}
.

Finished in 6.76 seconds (files took 0.13213 seconds to load)
1 example, 0 failures

        7.19 real         6.84 user         0.32 sys
           491335680  maximum resident set size
```

**After:**

```
$ /usr/bin/time -l rspec spec/zip_tricks/streamer_spec.rb:31
Run options: include {:locations=>{"./spec/zip_tricks/streamer_spec.rb"=>[31]}}
.

Finished in 0.19886 seconds (files took 0.11675 seconds to load)
1 example, 0 failures

        0.56 real         0.46 user         0.08 sys
            34746368  maximum resident set size
```