@felixbuenemann 2020-11-25
I have already seen a case were the 25% variance check was not enough on JRuby, although by a tiny margin. Something like rspec-retry could be a solution, though it is more of a bandaid.