@felixbuenemann 2020-11-25
@julik Personals I tend towards putting such slow and potentially flipping tests under a tag and only run it on demand.

This could also be done as a benchmark with benchmark-ips, which implements quite a bit more logic to weed out variance across different test sizes than what I can do.

Any attempts to make them more reliable end up spending even more cpu time, since it comes down to weeding out variance through more tries or larger iteration counts. – I wen't down the rabbit hole with this one and you quickly spend seconds on a single test, where the hole remaining test suite can run in 2.5s without this spec.

There is also the case here that the behaviour is more like "eventually linear", if you look at the distribution for the tested range it goes from something like 50% to 150% of the 2x linear target. 

What do you think? – It would also be interesting to known what the original intention for adding the test was.