@julik 2020-11-25
The original intention behind the test can be found in #14 - we had a situation where we inadvertently added a feature which would scan through the file list. As the file list would grow longer, adding a new file would scan the entire list again, which ultimately resulted in an "accidentally quadratic" situation. This started being very noticeable with even a few thousand files, and when this was fixed we decided that it is important to "codify" the fact that it is fixed (and should stay fixed) with a test.

Now, initially you will see that it also used a custom assertion performance-wise, but we went with complexity-assert for the following reason: for someone just getting into the code of the library, it would be conceptually easier to understand what the code does, as it calls into a well-defined, documented dependency. I do understand that there is value in having faster tests, and we could make this test skip under most CI runs and only keep it in, say, one of the CI passes.

But I don't know whether it is very valuable to brutally optimise the performance of tests, knowing that we will have to sacrifice some amount of readability/clarity while doing so. In the library itself such optimisations really are justified because the performance cost is paid by all users of the library (sometimes unknowingly), but in tests the priorities are not so clear-cut.

Can we try an approach where we isolate these "heavy-duty" tests under a single CI pass and only pay the runtime cost there? I think we should also be able to tweak the amount of items/iterations that complexity assert does.

@piotrmurach do you have suggestions maybe?