@piotrmurach 2020-11-25
Thanks for tagging me, even if by accident. Enjoyed reading the discussion so far. Agree with @julik that keeping a perf test to capture potential regressions is useful. Otherwise, how will you ensure that some refactoring is not going to cause a regression? I like the idea of grouping 'heavy-duty' tests together. I tend to keep them in a separate folder and run with `spec:perf`. Separate job or workflow to run these is a good way to resolve this, on one hand, having a snappy base test suite and a backup of perf tests as required to merge changes. 

If you want to use 'out-of-the-box' matchers then [rspec-benchamark complexity matcher](https://github.com/piotrmurach/rspec-benchmark#14-complexity) would be a good candidate. It allows you to pick the sampling rate and threshold. If that feels too 'heavy' then you can try writing your own `be_linear` matcher based on the [benchmark-trend](https://github.com/piotrmurach/benchmark-trend). Would be curious to know how it compares with your current way.