@felixbuenemann 2020-11-25
I will need to look into GitHub Actions documentation what would be good ways to trigger this. There are also scheduled test runs that could be used for this or a filter that for example only triggers on merge to master.